requirements.txt
djangae==0.9.7
Django==1.10
django-cors-middleware==1.3.1
django-filter==0.14.0
djangorestframework==3.4.4
djangorestframework-jwt==1.8.0
PyJWT==1.4.2
